wildcards, docs
git-svn-id: svn://localhost/trunk@960 3bdad9f7-c964-442a-868d-553b97b55ca0

+++ surfit/src/surfit/hist.cpp
@@ -88,7 +88,7 @@ bool d_hist::getMinMaxZ(REAL & minz, REAL & maxz) const
 	return false;
 };
 
-int d_hist::size() const {
+size_t d_hist::size() const {
 	return cnt;
 };
